Atualizar o ex13.py

+++ ex13.py
@@ -1,11 +1,11 @@
 # Receba a quantidade de alimento em quilos. Calcule e mostre quantos dias durará esse alimento sabendo que a pessoa consome 50g ao dia
 # declarar
-valorI: int=0
-dias: int=0
-gramas: int=0
+valorI: float=0
+dias: float=0
+gramas: float=0
 #inicio
 #recebe quantos quilos de alimento
-valorI = int(input('Digite quantos quilos de alimento: '))
+valorI = float(input('Digite quantos quilos de alimento: '))
 
 #converte quilos pars gramas
 gramas= 1000*valorI
